extracted education number
